# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "KORA+ Audio Pilot API",
    "version": "1.2.1",
    "description": "S1.2-03B contract-only Admin Security gate layered on the immutable S1.2-01 surface. No runtime implementation, migration, provider call, seed media or private media location is supplied by this contract."
  },
...
  "servers": [{ "url": "/", "description": "Relative contract root only" }],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginCustomer",
        "summary": "Verifies an international phone and password before issuing a login OTP challenge.",
        "x-kora-clients": ["mobile"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Password-verified login challenge accepted without disclosing account details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/ClientError" }
        }
      }
    },
...
    "/api/v1/auth/sessions/refresh": {
      "post": {
        "operationId": "refreshCustomerSession",
        "summary": "Rotates a single-use refresh token for its session family.",
        "x-kora-clients": ["mobile"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotated session tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionEnvelope"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ClientError" }
        }
      }
    },
    "/api/v1/auth/sessions/current": {
      "delete": {
        "operationId": "revokeCurrentCustomerSession",
        "summary": "Idempotently revokes the current device-bound session.",
        "x-kora-clients": ["mobile"],
        "security": [{ "customerBearer": [] }],
        "responses": {
          "204": { "description": "Session revoked or already absent." },
          "401": { "$ref": "#/components/responses/ClientError" }
        }
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "operationId": "loginAdmin",
        "summary": "Verifies admin credentials without disclosing account state.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PUBLIC",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "ADMIN_SECURITY_EVENT",
        "x-kora-rate-limit-profile": "PASSWORD",
        "x-kora-fetch-metadata-policy": "REJECT_CROSS_SITE_REQUIRE_SAME_ORIGIN",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-security-record-atomic": true,
        "x-kora-json-body-policy": "APPLICATION_JSON_ONLY",
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/AdminOrigin" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic credential result and bounded pre-authentication context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPreAuthEnvelope"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/AdminPreAuthSetCookie"
              },
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/RateLimitedError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
    "/api/v1/admin/auth/totp/enrollments": {
      "post": {
        "operationId": "createAdminTotpEnrollment",
        "summary": "Creates a bounded TOTP enrollment without exposing its seed.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PREAUTH",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG_ADMIN_RECOVERY_IF_SERVER_CONTEXT_PROVEN_ELSE_ADMIN_SECURITY_EVENT",
        "x-kora-audit-context-proof": "SERVER_VERIFIED_MFA_RECOVERY_OWNERSHIP_ACTOR_BINDING_STATE_AND_EXPIRY_CLIENT_INPUT_NEVER_SUFFICIENT",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-security-record-atomic": true,
        "x-kora-transactional-audit": true,
        "x-kora-idempotent": true,
        "x-kora-idempotent-replay": "REPLAY_SAME_RESPONSE_FOR_SAME_PAYLOAD_DIFFERENT_PAYLOAD_409",
        "security": [
          {
            "adminPreAuthCookie": [],
            "adminCsrfCookie": [],
            "adminCsrfHeader": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/AdminOrigin" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "201": {
            "description": "Enrollment metadata created; the seed remains server-side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTotpEnrollmentEnvelope"
                }
              }
            },
            "headers": {
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "403": { "$ref": "#/components/responses/ClientError" },
          "409": { "$ref": "#/components/responses/ClientError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
    "/api/v1/admin/auth/totp/enrollments/{enrollmentId}/qr": {
      "post": {
        "operationId": "deliverAdminTotpEnrollmentQr",
        "summary": "Delivers TOTP enrollment QR bytes exactly once.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PREAUTH",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG_ADMIN_RECOVERY_IF_SERVER_CONTEXT_PROVEN_ELSE_ADMIN_SECURITY_EVENT",
        "x-kora-audit-context-proof": "SERVER_VERIFIED_MFA_RECOVERY_OWNERSHIP_ACTOR_BINDING_STATE_AND_EXPIRY_CLIENT_INPUT_NEVER_SUFFICIENT",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-security-record-atomic": true,
        "x-kora-transactional-audit": true,
        "x-kora-idempotent": true,
        "x-kora-idempotent-replay": "SECRET_RESPONSE_RETRY_409",
        "x-kora-non-persistable-response": true,
        "x-kora-non-loggable-response": true,
        "security": [
          {
            "adminPreAuthCookie": [],
            "adminCsrfCookie": [],
            "adminCsrfHeader": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/EnrollmentId" },
          { "$ref": "#/components/parameters/AdminOrigin" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Single-delivery QR image; seed and provisioning URI never enter JSON, logs or audit payloads.",
            "headers": {
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            },
            "content": {
              "image/png": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ClientError" },
          "403": { "$ref": "#/components/responses/ClientError" },
          "409": { "$ref": "#/components/responses/ClientError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
    "/api/v1/admin/auth/totp/enrollments/{enrollmentId}/confirm": {
      "post": {
        "operationId": "confirmAdminTotpEnrollment",
        "summary": "Confirms TOTP and creates a session plus exactly ten recovery codes.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PREAUTH",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG",
        "x-kora-rate-limit-profile": "TOTP",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-transactional-audit": true,
        "x-kora-json-body-policy": "APPLICATION_JSON_ONLY",
        "x-kora-idempotent": true,
        "x-kora-idempotent-replay": "SECRET_RESPONSE_RETRY_409",
        "x-kora-non-persistable-response": true,
        "x-kora-non-loggable-response": true,
        "security": [
          {
            "adminPreAuthCookie": [],
            "adminCsrfCookie": [],
            "adminCsrfHeader": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/EnrollmentId" },
          { "$ref": "#/components/parameters/AdminOrigin" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTotpVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TOTP confirmed atomically with session and ten one-time codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEnrollmentConfirmationEnvelope"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/AdminRefreshSetCookie"
              },
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "403": { "$ref": "#/components/responses/ClientError" },
          "409": { "$ref": "#/components/responses/ClientError" },
          "410": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/RateLimitedError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
    "/api/v1/admin/auth/totp/verify": {
      "post": {
        "operationId": "verifyAdminTotp",
        "summary": "Completes every admin login with RFC 6238 verification.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PREAUTH",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG",
        "x-kora-rate-limit-profile": "TOTP",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-transactional-audit": true,
        "x-kora-json-body-policy": "APPLICATION_JSON_ONLY",
        "security": [
          {
            "adminPreAuthCookie": [],
            "adminCsrfCookie": [],
            "adminCsrfHeader": []
          }
        ],
        "parameters": [{ "$ref": "#/components/parameters/AdminOrigin" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTotpVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin session established.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionEnvelope"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/AdminRefreshSetCookie"
              },
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "410": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/RateLimitedError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
    "/api/v1/admin/auth/recovery-codes/verify": {
      "post": {
        "operationId": "verifyAdminRecoveryCode",
        "summary": "Consumes a recovery code and creates only a bounded MFA recovery context.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "PREAUTH",
        "x-kora-roles": [],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG",
        "x-kora-rate-limit-profile": "RECOVERY",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-transactional-audit": true,
        "x-kora-json-body-policy": "APPLICATION_JSON_ONLY",
        "security": [
          {
            "adminPreAuthCookie": [],
            "adminCsrfCookie": [],
            "adminCsrfHeader": []
          }
        ],
        "parameters": [{ "$ref": "#/components/parameters/AdminOrigin" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRecoveryCodeVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recovery code consumed, old sessions revoked, no session created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRecoveryContextEnvelope"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/AdminPreAuthSetCookie"
              },
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/RateLimitedError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
...
    "/api/v1/admin/auth/step-up": {
      "post": {
        "operationId": "stepUpAdminSession",
        "summary": "Refreshes sensitive-action proof for five minutes.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "ADMIN_SESSION",
        "x-kora-roles": [
          "SUPER_ADMIN",
          "CONTENT_EDITOR",
          "FINANCE_MANAGER",
          "SUPPORT"
        ],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG",
        "x-kora-rate-limit-profile": "TOTP",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-transactional-audit": true,
        "x-kora-json-body-policy": "APPLICATION_JSON_ONLY",
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminOrigin" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStepUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Step-up proof established for five minutes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStepUpEnvelope"
                }
              }
            },
            "headers": {
              "Cache-Control": { "$ref": "#/components/headers/NoStore" },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/ClientError" },
          "403": { "$ref": "#/components/responses/ClientError" },
          "429": { "$ref": "#/components/responses/RateLimitedError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
...
    "/api/v1/admin/auth/sessions/current": {
      "delete": {
        "operationId": "revokeCurrentAdminSession",
        "summary": "Revokes the current admin session family.",
        "x-kora-clients": ["admin-security"],
        "x-kora-delivery-slice": "S1.2-03C1",
        "x-kora-auth-class": "ADMIN_SESSION",
        "x-kora-roles": [
          "SUPER_ADMIN",
          "CONTENT_EDITOR",
          "FINANCE_MANAGER",
          "SUPPORT"
        ],
        "x-kora-step-up-required": false,
        "x-kora-audit-sink": "AUDIT_LOG",
        "x-kora-origin-policy": "EXACT_ALLOWLIST",
        "x-kora-transactional-audit": true,
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminOrigin" }],
        "responses": {
          "204": { "description": "Current session revoked." },
          "401": { "$ref": "#/components/responses/ClientError" },
          "503": { "$ref": "#/components/responses/AdminServiceUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "customerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "RS256 JWT; 15 minute maximum",
        "description": "Short-lived admin access credential. Runtime authorization must reload active session, account status and authorizationVersion; JWT role claims are never sole authority."
      },
      "adminPreAuthCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kora_admin_preauth",
        "description": "HttpOnly Secure SameSite=Strict host-only bounded pre-authentication context; never an admin session."
      },
      "adminRefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kora_admin_refresh",
        "description": "HttpOnly Secure SameSite=Strict host-only opaque 256-bit one-shot refresh token."
      },
      "adminCsrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kora_admin_csrf",
        "description": "Secure SameSite=Strict host-only CSRF token cookie."
      },
      "adminCsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kora-CSRF",
        "description": "Exact double-submit value matching adminCsrfCookie."
      },
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kora-Sandbox-Signature"
      },
...
  }
}